+++ resources/views/products/create.blade.php
@@ -415,10 +415,7 @@
                             </div>
                         </div>
 
-                    </div>
-
-                    <!-- Image Upload Section -->
-                    <div class="xl:col-span-2">
+                        <!-- Image Upload Section -->
                         <div class="bg-white dark:bg-gray-800 shadow rounded-lg p-6">
                             <div class="border-b border-gray-200 dark:border-gray-700 pb-3 mb-4">
                                 <h3 class="text-lg font-semibold text-gray-900 dark:text-gray-100 flex items-center">
@@ -455,6 +452,7 @@
                                 </div>
                             </div>
                         </div>
+
                     </div>
 
                     <!-- Right Column -->